requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
pypdf
python-docx